Remove algolia stuff from main.js Rename custom explore to categories Remove edits to settings.vue Remove nodemon
