+++ package.json
@@ -16,7 +16,6 @@
     "instantsearch.css": "^7.4.2",
     "lodash.clonedeep": "^4.5.0",
     "moment": "^2.25.3",
-    "nodemon": "^2.0.4",
     "register-service-worker": "^1.7.1",
     "vue": "^2.6.11",
     "vue-instantsearch": "^2.7.0",
@@ -38,7 +37,7 @@
     "eslint-plugin-prettier": "^3.1.3",
     "eslint-plugin-vue": "^6.2.2",
     "husky": "^4.2.5",
-    "lint-staged": "^10.2.2",
+    "lint-staged": "^10.2.6",
     "prettier": "^1.19.1",
     "sass": "^1.26.5",
     "sass-loader": "^8.0.2",
